+++ modules/schema/src/main/resources/lucuma/odb/graphql/OdbSchema.graphql
