+++ modules/schema/src/main/scala/lucuma/odb/json/all.scala
@@ -1,0 +1,10 @@
+// Copyright (c) 2016-2023 Association of Universities for Research in Astronomy, Inc. (AURA)
+// For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
+
+package lucuma.odb.json
+
+object all
+  extends AngleCodec
+     with NumericCodecs
+     with SourceProfileCodec
+     with WavelengthCodec
