+++ modules/schema/src/test/scala/lucuma/odb/json/SourceProfileCodecSuite.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2016-2023 Association of Universities for Research in Astronomy, Inc. (AURA)
 // For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-package lucuma.odb.graphql.instances
+package lucuma.odb.json
 
 import io.circe.testing.ArbitraryInstances
 import io.circe.testing.CodecTests
@@ -12,7 +12,7 @@ import munit.DisciplineSuite
 class SourceProfileCodecSuite extends DisciplineSuite with ArbitraryInstances {
 
   import ArbSourceProfile.given
-  import SourceProfileCodec.given
+  import sourceprofile.given
 
   checkAll("SourceProfileCodec", CodecTests[SourceProfile].codec)
