+++ modules/service/src/main/scala/lucuma/odb/service/ItcInputService.scala
@@ -35,7 +35,7 @@ import lucuma.core.model.User
 import lucuma.itc.client.InstrumentMode
 import lucuma.itc.client.SpectroscopyModeInput
 import lucuma.odb.data.ObservingModeType
-import lucuma.odb.graphql.instances.SourceProfileCodec.given
+import lucuma.odb.json.sourceprofile.given
 import lucuma.odb.util.Codecs.*
 import skunk.*
 import skunk.circe.codec.json.*
